created sub-package that handles API-interation & moved logic
