+++ pkg/venafi/cloud/cloud_api/cloud_structs/edgeencryptionkeys.go
@@ -14,15 +14,9 @@
  * limitations under the License.
  */
 
-package cloud
+package cloud_structs
 
-type team struct {
-	ID        string `json:"id"`
-	Name      string `json:"name"`
-	Role      string `json:"role"`
-	CompanyID string `json:"companyId"`
-}
-
-type teams struct {
-	Teams []team `json:"teams"`
+// RESP: GET outagedetection/v1/edgeencryptionkeys/%s
+type EdgeEncryptionKey struct {
+	Key string `json:"key,omitempty"`
 }
