+++ pkg/venafi/cloud/cloud_api/error_test.go
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Venafi, Inc.
+ * Copyright 2022 Venafi, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,35 +14,40 @@
  * limitations under the License.
  */
 
-package cloud
+package cloud_api
 
 import (
+	"bytes"
+	"errors"
 	"testing"
 )
 
 func TestParseResponseErrors(t *testing.T) {
 	data := []byte("{\"errors\":[{\"code\":10128,\"message\":\"Invalid change in apiKey status\",\"args\":[]}]}")
-	errors, err := parseResponseErrors(data)
-	if err != nil {
-		t.Fatalf("err is not nil, err: %s", err)
-	}
+	err := parseResponseErrors(500, bytes.NewBuffer(data))
 
-	if len(errors) != 1 {
-		t.Fatalf("Parsed error count was not what was expected.  Expected: 1 Actual: %d", len(errors))
+	if target := new(StructuredResponseBody); errors.As(err, target) {
+		if len(target.Body.Errors) != 1 {
+			t.Fatalf("ParseResponseErrors returned incorrect error count.  Expected: 1 Actual: %d", len(target.Body.Errors))
+		}
+	} else {
+		t.Fatalf("Wrong error type expected %T, got %T instead", target, err)
 	}
 }
 
 func TestParseResponseErrorWithArgs(t *testing.T) {
 	data := []byte("{\"errors\": [{\"code\": 10726,\"message\": \"Distinguished name component CN with value \\\"test.venafi.io\\\" is invalid\",\"args\": [\"CN\",\"test.venafi.io\",[\".*.example.com\",\".*.example.org\",\".*.example.net\",\".*.invalid\",\".*.local\",\".*.localhost\",\".*.test\"]]}]}")
-	errors, err := parseResponseErrors(data)
-	if err != nil {
-		t.Fatalf("err is not nil, err: %s", err)
-	}
+	err := parseResponseErrors(500, bytes.NewBuffer(data))
 
-	if len(errors) != 1 {
-		t.Fatalf("ParseResponseErrors returned incorrect error count.  Expected: 1 Actual: %d", len(errors))
-	}
-	if errors[0].Code != 10726 {
-		t.Fatalf("ParseResponseErrors returned incorrect code.  Expected: 10726 Actual: %d", errors[0].Code)
+	if target := new(StructuredResponseBody); errors.As(err, target) {
+		if len(target.Body.Errors) != 1 {
+			t.Fatalf("ParseResponseErrors returned incorrect error count.  Expected: 1 Actual: %d", len(target.Body.Errors))
+		}
+
+		if target.Body.Errors[0].Code != 10726 {
+			t.Fatalf("ParseResponseErrors returned incorrect code.  Expected: 10726 Actual: %d", target.Body.Errors[0].Code)
+		}
+	} else {
+		t.Fatalf("Wrong error type expected %T, got %T instead", target, err)
 	}
 }
