+++ pkg/venafi/cloud/search.go
@@ -17,76 +17,16 @@
 package cloud
 
 import (
-	"encoding/json"
-	"fmt"
-	"github.com/Venafi/vcert/v4/pkg/certificate"
 	"log"
 	"math"
-	"net/http"
 	"strings"
 	"time"
+
+	"github.com/Venafi/vcert/v4/pkg/certificate"
+	"github.com/Venafi/vcert/v4/pkg/venafi/cloud/cloud_api/cloud_structs"
 )
 
-type SearchRequest struct {
-	Expression *Expression  `json:"expression"`
-	Ordering   *interface{} `json:"ordering,omitempty"`
-	Paging     *Paging      `json:"paging,omitempty"`
-	// ordering is not used here so far
-	// "ordering": {"orders": [{"direction": "ASC", "field": "subjectCN"},{"direction": "DESC", "field": "keyStrength"}]},
-}
-
-type Expression struct {
-	Operator Operator  `json:"operator,omitempty"`
-	Operands []Operand `json:"operands,omitempty"`
-}
-
-type Operand struct {
-	Field    Field       `json:"field"`
-	Operator Operator    `json:"operator"`
-	Value    interface{} `json:"value,omitempty"`
-	Values   interface{} `json:"values,omitempty"`
-}
-
-type Field string
-type Operator string
-
-type Paging struct {
-	PageNumber int `json:"pageNumber"`
-	PageSize   int `json:"pageSize"`
-}
-
-const (
-	EQ    Operator = "EQ"
-	FIND  Operator = "FIND"
-	GT    Operator = "GT"
-	GTE   Operator = "GTE"
-	IN    Operator = "IN"
-	LT    Operator = "LT"
-	LTE   Operator = "LTE"
-	MATCH Operator = "MATCH"
-	AND   Operator = "AND"
-)
-
-type CertificateSearchResponse struct {
-	Count        int           `json:"count"`
-	Certificates []Certificate `json:"certificates"`
-}
-
-type Certificate struct {
-	Id                            string              `json:"id"`
-	ManagedCertificateId          string              `json:"managedCertificateId"`
-	CertificateRequestId          string              `json:"certificateRequestId"`
-	SubjectCN                     []string            `json:"subjectCN"`
-	SubjectAlternativeNamesByType map[string][]string `json:"subjectAlternativeNamesByType"`
-	SerialNumber                  string              `json:"serialNumber"`
-	Fingerprint                   string              `json:"fingerprint"`
-	ValidityStart                 string              `json:"validityStart"`
-	ValidityEnd                   string              `json:"validityEnd"`
-	ApplicationIds                []string            `json:"applicationIds"`
-	/* ... and many more fields ... */
-}
-
-func (c Certificate) ToCertificateInfo() certificate.CertificateInfo {
+func certificateToCertificateInfo(c *cloud_structs.Certificate) certificate.CertificateInfo {
 	var cn string
 	if len(c.SubjectCN) > 0 {
 		cn = c.SubjectCN[0]
@@ -120,30 +60,6 @@ func (c Certificate) ToCertificateInfo() certificate.CertificateInfo {
 	}
 }
 
-func ParseCertificateSearchResponse(httpStatusCode int, body []byte) (searchResult *CertificateSearchResponse, err error) {
-	switch httpStatusCode {
-	case http.StatusOK:
-		var searchResult = &CertificateSearchResponse{}
-		err = json.Unmarshal(body, searchResult)
-		if err != nil {
-			return nil, fmt.Errorf("failed to parse search results: %s, body: %s", err, body)
-		}
-		return searchResult, nil
-	default:
-		if body != nil {
-			respErrors, err := parseResponseErrors(body)
-			if err == nil {
-				respError := fmt.Sprintf("Unexpected status code on Venafi Cloud certificate search. Status: %d\n", httpStatusCode)
-				for _, e := range respErrors {
-					respError += fmt.Sprintf("Error Code: %d Error: %s\n", e.Code, e.Message)
-				}
-				return nil, fmt.Errorf(respError)
-			}
-		}
-		return nil, fmt.Errorf("unexpected status code on Venafi Cloud certificate search. Status: %d", httpStatusCode)
-	}
-}
-
 // returns everything up to the last slash (if any)
 //
 // example:
@@ -168,18 +84,18 @@ func getAppNameFromZone(zone string) string {
 }
 
 // TODO: test this function
-func formatSearchCertificateArguments(cn string, sans *certificate.Sans, certMinTimeLeft time.Duration) *SearchRequest {
+func formatSearchCertificateArguments(cn string, sans *certificate.Sans, certMinTimeLeft time.Duration) *cloud_structs.SearchRequest {
 	// convert a time.Duration to days
 	certMinTimeDays := math.Floor(certMinTimeLeft.Hours() / 24)
 
 	// generate base request
-	req := &SearchRequest{
-		Expression: &Expression{
-			Operator: AND,
-			Operands: []Operand{
+	req := &cloud_structs.SearchRequest{
+		Expression: &cloud_structs.Expression{
+			Operator: cloud_structs.AND,
+			Operands: []cloud_structs.Operand{
 				{
 					Field:    "validityPeriodDays",
-					Operator: GTE,
+					Operator: cloud_structs.GTE,
 					Value:    certMinTimeDays,
 				},
 			},
@@ -187,18 +103,18 @@ func formatSearchCertificateArguments(cn string, sans *certificate.Sans, certMin
 	}
 
 	if sans != nil && sans.DNS != nil {
-		addOperand(req, Operand{
+		addOperand(req, cloud_structs.Operand{
 			Field:    "subjectAlternativeNameDns",
-			Operator: IN,
+			Operator: cloud_structs.IN,
 			Values:   sans.DNS,
 		})
 	}
 
 	// only if a CN is provided, we add the field to the search request
 	if cn != "" {
-		addOperand(req, Operand{
+		addOperand(req, cloud_structs.Operand{
 			Field:    "subjectCN",
-			Operator: EQ,
+			Operator: cloud_structs.EQ,
 			Value:    cn,
 		})
 	}
@@ -206,7 +122,7 @@ func formatSearchCertificateArguments(cn string, sans *certificate.Sans, certMin
 	return req
 }
 
-func addOperand(req *SearchRequest, o Operand) *SearchRequest {
+func addOperand(req *cloud_structs.SearchRequest, o cloud_structs.Operand) *cloud_structs.SearchRequest {
 	req.Expression.Operands = append(req.Expression.Operands, o)
 	return req
 }
