+++ pkg/venafi/cloud/search_test.go
@@ -18,26 +18,28 @@ package cloud
 
 import (
 	"encoding/json"
-	"github.com/Venafi/vcert/v4/pkg/certificate"
-	"github.com/Venafi/vcert/v4/pkg/util"
 	"testing"
 	"time"
+
+	"github.com/Venafi/vcert/v4/pkg/certificate"
+	"github.com/Venafi/vcert/v4/pkg/util"
+	"github.com/Venafi/vcert/v4/pkg/venafi/cloud/cloud_api/cloud_structs"
 )
 
 func TestSearchRequest(t *testing.T) {
 
 	// encoding to JSON
-	req := &SearchRequest{
-		Expression: &Expression{
-			Operands: []Operand{
+	req := &cloud_structs.SearchRequest{
+		Expression: &cloud_structs.Expression{
+			Operands: []cloud_structs.Operand{
 				{
 					Field:    "fingerprint",
-					Operator: MATCH,
+					Operator: cloud_structs.MATCH,
 					Value:    "A7BDECDA0B67D5CEF28D6C8C7D7CFA882E3DC9D6",
 				},
 			},
 		},
-		Paging: &Paging{10, 10},
+		Paging: &cloud_structs.Paging{10, 10},
 	}
 	var expectedJson = `{"expression":{"operands":[{"field":"fingerprint","operator":"MATCH","value":"A7BDECDA0B67D5CEF28D6C8C7D7CFA882E3DC9D6"}]},"paging":{"pageNumber":10,"pageSize":10}}`
 
@@ -51,7 +53,7 @@ func TestSearchRequest(t *testing.T) {
 	t.Logf("%s\n", data)
 
 	// decoding from JSON
-	var req2 = &SearchRequest{}
+	var req2 = &cloud_structs.SearchRequest{}
 	err = json.Unmarshal(data, req2)
 	if err != nil {
 		t.Fatal(err)
@@ -66,12 +68,12 @@ func TestSearchRequest(t *testing.T) {
 	}
 
 	// if Paging is not specified, it should not be included to JSON
-	req = &SearchRequest{
-		Expression: &Expression{
-			Operands: []Operand{
+	req = &cloud_structs.SearchRequest{
+		Expression: &cloud_structs.Expression{
+			Operands: []cloud_structs.Operand{
 				{
 					Field:    "fingerprint",
-					Operator: MATCH,
+					Operator: cloud_structs.MATCH,
 					Value:    "A7BDECDA0B67D5CEF28D6C8C7D7CFA882E3DC9D6",
 				},
 			},
@@ -85,166 +87,6 @@ func TestSearchRequest(t *testing.T) {
 	}
 	if string(data) != expectedJson {
 		t.Fatalf("expected different JSON:\nhave:     %s\nexpected: %s", data, expectedJson)
-	}
-}
-
-func TestParseCertificateSearchResponse(t *testing.T) {
-	var code int
-	var body []byte
-	var searchResult *CertificateSearchResponse
-	var err error
-
-	code = 200
-	body = []byte(`
-{
-  "count": 1,
-  "certificates": [
-    {
-      "id": "ab239880-5de9-11e8-bb9b-8d6e819a14f1",
-      "companyId": "b5ed6d60-22c4-11e7-ac27-035f0608fd2c",
-      "managedCertificateId": "ab239881-5de9-11e8-bb9b-8d6e819a14f1",
-      "fingerprint": "73CF2CC98C7DEC4045EDB93151750F5B9609FF44",
-      "issuerCertificateIds": [
-        "828a2b70-22ce-11e7-ba19-0da4a5ff6335",
-        "82734810-22ce-11e7-ba19-0da4a5ff6335"
-      ],
-      "certificateRequestId": "8ceb8ad0-5de9-11e8-9c7a-e596bbf80f56",
-      "certificateSource": "USER_PROVIDED",
-      "certificateStatuses": [
-        "NONE"
-      ],
-      "certificateType": "END_ENTITY",
-      "ownerUsername": "alexander.tarasenko@venafi.com",
-      "creationDate": "2018-05-22T17:58:01.480+00:00",
-      "modificationDate": "2018-05-22T17:58:01.480+00:00",
-      "totalInstanceCount": 0,
-      "validityStart": "2018-05-22T00:00:00.000+00:00",
-      "validityEnd": "2018-08-20T12:00:00.000+00:00",
-      "validityPeriodDays": 90,
-      "validityPeriodRange": "GT_30_DAYS_LTE_2_YEARS",
-      "selfSigned": false,
-      "signatureAlgorithm": "SHA256_WITH_RSA_ENCRYPTION",
-      "signatureHashAlgorithm": "SHA256",
-      "encryptionType": "RSA",
-      "keyStrength": 2048,
-      "publicKeyHash": "0048AA1D7E2F0017F9CA2E687D8776A1A340553D",
-      "subjectKeyIdentifierHash": "C6E7C18CADE684CB420CA4764A6469086536D08E",
-      "authorityKeyIdentifierHash": "AC90A22B9320CE93369173BC3074121005D7F909",
-      "serialNumber": "07F3FE39F4E1A4B6075633ECFB748D84",
-      "subjectCN": [
-        "renew-test.venafi.example.com"
-      ],
-      "subjectOU": [
-        "SerialNumber"
-      ],
-      "subjectST": "California",
-      "subjectL": "Palo Alto",
-      "subjectC": "US",
-      "subjectAlternativeNamesByType": {
-        "otherName": [],
-        "rfc822Name": [],
-        "dNSName": [
-          "renew-test.venafi.example.com"
-        ],
-        "x400Address": [],
-        "directoryName": [],
-        "ediPartyName": [],
-        "uniformResourceIdentifier": [],
-        "iPAddress": [],
-        "registeredID": []
-      },
-      "subjectAlternativeNameDns": [
-        "renew-test.venafi.example.com"
-      ],
-      "issuerCN": [
-        "DigiCert Test SHA2 Intermediate CA-1"
-      ],
-      "issuerC": "US",
-      "keyUsage": [
-        "digitalSignature",
-        "keyEncipherment"
-      ],
-      "ocspNoCheck": false,
-      "compliance": {
-        "score": 0.8728395061728398
-      },
-      "instances": [
-        {
-	  	"id": "ab28c8a0-5de9-11e8-bb9b-8d6e819a14f1",
-	  	"certificateId": "ab239880-5de9-11e8-bb9b-8d6e819a14f1",
-	  	"managedCertificateId": "ab239881-5de9-11e8-bb9b-8d6e819a14f1",
-	  	"companyId": "b5ed6d60-22c4-11e7-ac27-035f0608fd2c",
-	  	"zoneId": "b5f69520-22c4-11e7-ac27-035f0608fd2c",
-	  	"fingerprint": "73CF2CC98C7DEC4045EDB93151750F5B9609FF44",
-	  	"certificateSource": "USER_PROVIDED",
-	  	"certificateStatuses": [
-	  		"NONE"
-	  	],
-	  	"ownerUsername": "alexander.tarasenko@venafi.com",
-	  	"creationDate": "2018-05-22T17:58:01.514+00:00",
-	  	"modificationDate": "2018-05-22T17:58:01.514+00:00",
-	  	"ipAddress": "254.254.254.254",
-	  	"ipAddressAsLong": 4278124286,
-	  	"hostname": " ",
-	  	"port": -1,
-	  	"sslProtocolsSecurityStatus": "UNKNOWN",
-	  	"cipherSuitesSecurityStatus": "UNKNOWN",
-	  	"compliance": {
-	  		"score": 0.0
-	  	}
-        }
-      ],
-      "applicationIds": []
-    }
-  ]
-}
-`)
-
-	searchResult, err = ParseCertificateSearchResponse(code, body)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if searchResult.Count != 1 {
-		t.Fatal("wrong count field")
-	}
-	if len(searchResult.Certificates) != 1 {
-		t.Fatal("wrong count")
-	}
-	if searchResult.Certificates[0].ManagedCertificateId != "ab239881-5de9-11e8-bb9b-8d6e819a14f1" {
-		t.Fatal("wrong ManagedCertificateId value")
-	}
-
-	code = 400
-	body = []byte("")
-	searchResult, err = ParseCertificateSearchResponse(code, body)
-	if err == nil {
-		t.Fatal("should trigger error")
-	}
-
-	code = 400
-	body = []byte(`
-		{
-		  "errors": [
-		    {
-		      "code": 1004,
-		      "message": "Invalid or missing request header [SESSION, tppl-api-key]",
-		      "args": [
-		        [
-		          "SESSION",
-		          "tppl-api-key"
-		        ]
-		      ]
-		    },
-		    {
-		      "code": 1005,
-		      "message": "Something else was wrong"
-		    }
-		  ]
-		}
-	`)
-	searchResult, err = ParseCertificateSearchResponse(code, body)
-	if err == nil {
-		t.Fatal("JSON body should trigger error")
 	}
 }
 
@@ -296,20 +138,20 @@ func TestFormatSearchCertificateArguments(t *testing.T) {
 	testCases := []struct {
 		name     string
 		input    FormatSearchCertificateArgumentsMock
-		expected *SearchRequest
+		expected *cloud_structs.SearchRequest
 	}{
 		{
 			// test empty arguments, should return just the validityPeriodDays
 			// argument
 			name:  "Empty",
 			input: FormatSearchCertificateArgumentsMock{},
-			expected: &SearchRequest{
-				Expression: &Expression{
-					Operator: AND,
-					Operands: []Operand{
+			expected: &cloud_structs.SearchRequest{
+				Expression: &cloud_structs.Expression{
+					Operator: cloud_structs.AND,
+					Operands: []cloud_structs.Operand{
 						{
 							Field:    "validityPeriodDays",
-							Operator: GTE,
+							Operator: cloud_structs.GTE,
 							Value:    0,
 						},
 					},
@@ -323,18 +165,18 @@ func TestFormatSearchCertificateArguments(t *testing.T) {
 			input: FormatSearchCertificateArgumentsMock{
 				cn: "test.example.com",
 			},
-			expected: &SearchRequest{
-				Expression: &Expression{
-					Operator: AND,
-					Operands: []Operand{
+			expected: &cloud_structs.SearchRequest{
+				Expression: &cloud_structs.Expression{
+					Operator: cloud_structs.AND,
+					Operands: []cloud_structs.Operand{
 						{
 							Field:    "validityPeriodDays",
-							Operator: GTE,
+							Operator: cloud_structs.GTE,
 							Value:    0,
 						},
 						{
 							Field:    "subjectCN",
-							Operator: EQ,
+							Operator: cloud_structs.EQ,
 							Value:    "test.example.com",
 						},
 					},
@@ -348,18 +190,18 @@ func TestFormatSearchCertificateArguments(t *testing.T) {
 			input: FormatSearchCertificateArgumentsMock{
 				sans: &certificate.Sans{DNS: []string{"one.example.com"}},
 			},
-			expected: &SearchRequest{
-				Expression: &Expression{
-					Operator: AND,
-					Operands: []Operand{
+			expected: &cloud_structs.SearchRequest{
+				Expression: &cloud_structs.Expression{
+					Operator: cloud_structs.AND,
+					Operands: []cloud_structs.Operand{
 						{
 							Field:    "validityPeriodDays",
-							Operator: GTE,
+							Operator: cloud_structs.GTE,
 							Value:    0,
 						},
 						{
 							Field:    "subjectAlternativeNameDns",
-							Operator: IN,
+							Operator: cloud_structs.IN,
 							Values:   []string{"one.example.com"},
 						},
 					},
@@ -373,18 +215,18 @@ func TestFormatSearchCertificateArguments(t *testing.T) {
 			input: FormatSearchCertificateArgumentsMock{
 				sans: &certificate.Sans{DNS: []string{"one.example.com", "two.example.com"}},
 			},
-			expected: &SearchRequest{
-				Expression: &Expression{
-					Operator: AND,
-					Operands: []Operand{
+			expected: &cloud_structs.SearchRequest{
+				Expression: &cloud_structs.Expression{
+					Operator: cloud_structs.AND,
+					Operands: []cloud_structs.Operand{
 						{
 							Field:    "validityPeriodDays",
-							Operator: GTE,
+							Operator: cloud_structs.GTE,
 							Value:    0,
 						},
 						{
 							Field:    "subjectAlternativeNameDns",
-							Operator: IN,
+							Operator: cloud_structs.IN,
 							Values:   []string{"one.example.com", "two.example.com"},
 						},
 					},
@@ -399,23 +241,23 @@ func TestFormatSearchCertificateArguments(t *testing.T) {
 				cn:   "one.example.com",
 				sans: &certificate.Sans{DNS: []string{"one.example.com"}},
 			},
-			expected: &SearchRequest{
-				Expression: &Expression{
-					Operator: AND,
-					Operands: []Operand{
+			expected: &cloud_structs.SearchRequest{
+				Expression: &cloud_structs.Expression{
+					Operator: cloud_structs.AND,
+					Operands: []cloud_structs.Operand{
 						{
 							Field:    "validityPeriodDays",
-							Operator: GTE,
+							Operator: cloud_structs.GTE,
 							Value:    0,
 						},
 						{
 							Field:    "subjectAlternativeNameDns",
-							Operator: IN,
+							Operator: cloud_structs.IN,
 							Values:   []string{"one.example.com"},
 						},
 						{
 							Field:    "subjectCN",
-							Operator: EQ,
+							Operator: cloud_structs.EQ,
 							Value:    "one.example.com",
 						},
 					},
@@ -430,23 +272,23 @@ func TestFormatSearchCertificateArguments(t *testing.T) {
 				cn:   "one.example.com",
 				sans: &certificate.Sans{DNS: []string{"one.example.com", "two.example.com"}},
 			},
-			expected: &SearchRequest{
-				Expression: &Expression{
-					Operator: AND,
-					Operands: []Operand{
+			expected: &cloud_structs.SearchRequest{
+				Expression: &cloud_structs.Expression{
+					Operator: cloud_structs.AND,
+					Operands: []cloud_structs.Operand{
 						{
 							Field:    "validityPeriodDays",
-							Operator: GTE,
+							Operator: cloud_structs.GTE,
 							Value:    0,
 						},
 						{
 							Field:    "subjectAlternativeNameDns",
-							Operator: IN,
+							Operator: cloud_structs.IN,
 							Values:   []string{"one.example.com", "two.example.com"},
 						},
 						{
 							Field:    "subjectCN",
-							Operator: EQ,
+							Operator: cloud_structs.EQ,
 							Value:    "one.example.com",
 						},
 					},
